Content
No need for timeline.
We're naming responsibilities in the introductions for the subsystems, so maybe get rid of responsbilities/accomplishments section.

The Syncing subsystem needs the diagram used in the final presentation.  Priority algorithm subsystems needs the first table that was used in the final presentation. 

Training needs to be altered. Although I'm not sure exactly what to put there considerig we have a seperate manual already. Maybe simply not have that part?

Notifications: Crop out the notification segment of the notification demonstration from the final presentation, display that there.